fix(frontend): keep lib/actions/sql out of client bundles
executeQuery now resolves the session actor through lib/auth-session
(next/headers), so client components can no longer reach it transitively.
Cluster components imported UNCLUSTERED_ID/EventCluster from
lib/actions/clusters/index.ts, which pulls executeQuery in; those live in
a client-safe lib/actions/clusters/types.ts now. Fixes the webpack build.
