+++ frontend/components/landing/sections/has-this-issue/clusters-chart.tsx
@@ -7,7 +7,8 @@ import { numberFormatter, selectNiceTicksFromData } from "@/components/chart-bui
 import { type TimeSeriesChartConfig, type TimeSeriesDataPoint } from "@/components/charts/time-series-chart/types";
 import { getTickCountForWidth } from "@/components/charts/time-series-chart/utils";
 import { ChartContainer } from "@/components/ui/chart";
-import { type ClusterStatsDataPoint, type EventCluster } from "@/lib/actions/clusters";
+import { type ClusterStatsDataPoint } from "@/lib/actions/clusters";
+import { type EventCluster } from "@/lib/actions/clusters/types";
 import { withOpacity } from "@/lib/clusters/colors";
 
 // Landing copy of components/signal/clusters-section/cluster-stacked-chart and
